refactor: adding extensions folder
