+++ cinema-planner/cinema-planner/Extensions/CollectionViewTableViewCell+Extension.swift
@@ -1,0 +1,21 @@
+//
+//  CollectionViewTableViewCell+Extension.swift
+//  cinema-planner
+//
+//  Created by Mpilo Pillz on 2022/11/20.
+//
+
+import UIKit
+
+extension CollectionViewTableViewCell: UICollectionViewDelegate, UICollectionViewDataSource {
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "cell", for: indexPath)
+        cell.backgroundColor = .green
+        return cell
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 10
+    }
+}
